nq.c: use mkstemp(3) to create temporary file
In order to open and lock the `lockfile' atomically, a file name with
prefix `.' was used, and renameat(2) was called later.

In this commit, the temporary file is instead created by mkstemp(3).

Caution: Since the dot prefix is no longer present, `lockfile+1' should
be replaced by `lockfile' in later commits.