Avoid unnecessarily overwriting Makefile variables
The Makefile in its original form sets certain variables, like PREFIX, in a way which will override the user's (or build environment's) existing settings. Switching the "=" to the "?=" symbol allows the host environment to set things like the compiler and PREFIX automatically, without needing to patch the Makefile.

This will make porting to other platforms like FreeBSD easier.